+++ CHANGES.md
@@ -31,6 +31,7 @@ Release Notes.
 * Bump up swck dependency for transitive dep upgrade by @kezhenxu94 in https://github.com/apache/skywalking-cli/pull/162
 * Add the sub-commands for query sorted metrics/records by @mrproliu in https://github.com/apache/skywalking-cli/pull/163
 * Add compatibility documentation by @mrproliu in https://github.com/apache/skywalking-cli/pull/164
+* Add the sub-command `records list` for adapt the new record query API by @mrproliu in https://github.com/apache/skywalking-cli/pull/167
 
 0.10.0
 ------------------